FIXUP for review
 * Move tests to rebasehelper/tests/test_specfile.py
 * Handle locale.Error
